refactor: split Indexable interface - `Identifiable` only contains id related accessors - `Indexable` now extends `Identifiable` and contains accessors for `referenceId` and `referenceType` https://gravitee.atlassian.net/browse/APIM-2320
